refactor: :fire: Moved `sc-owner` unit test to `host` crate
