+++ host/Cargo.toml
@@ -13,3 +13,10 @@ primitive-types = { version = "0.12.0", default-features = false }
 tokio = { version = "1", features = ["full"] }
 dotenv = "0.15"
 bytemuck = "1.14"
+hex = { version = "0.4", default-features = false, features = ["alloc"] }
+alloy-sol-types = { version = "1.0.0", features = ["json"] }
+alloy-provider = { version = "0.14.0" }
+alloy-signer-local = { version = "0.14.0" }
+alloy-contract = { version = "0.14.0" }
+url = { version = "2" }
+anyhow = "1.0"
