+++ sc-owner/Cargo.toml
@@ -10,9 +10,3 @@ dotenv = "0.15"
 serde = { version = "1.0", default-features = false, features = ["derive"] }
 serde_json = { version = "1.0", default-features = false, features = ["alloc"] }
 hex = { version = "0.4", default-features = false, features = ["alloc"] }
-alloy-sol-types = { version = "1.0.0", features = ["json"] }
-alloy-provider = { version = "0.14.0" }
-alloy-signer-local = { version = "0.14.0" }
-alloy-contract = { version = "0.14.0" }
-url = { version = "2" }
-anyhow = "1.0"
